weight-service/requirements.txt
paho-mqtt
mysql-connector-python>=8.0
flask
flask-cors
